creating MQTTEvents enumeration instead of individual string names. consolidating into a single file to eliminate import errors
